Improved findBestMoves (theoretical) performance
By checking the time only every 1000, 10,000, etc. iterations, we save a
huge amount of time and can thus run the loop for longer. However, the less
frequently we check the time, the larger possible delay we'll have before
timing out (could take timeLimit + some number of milliseconds). The
maximum delay should be the time it takes to do the number of iterations
equal to the modulus used on count.