Rewrote findMoves algorithm; 2500x speed increase
Rewrote the findMoves algorithm to take advantage of the TST structure. A
sudden insight that the TST's space-efficiency comes from storing only the
nodes that actually exist - so by searching for lots of nodes that didn't
exist (particularly because we have to go left and right, instead of using
a direct array access like in a Trie) we waste tons of time. The new
algorithm uses an entirely different approach - because the TST represents
only the nodes that exist, we simply traverse the TST, following paths
that are possible with the pool of tiles we have remaining. By translating
the board into an array which represents the alphabet and subarrays containing
the positions of each letter in the board, we're able to traverse the TST by
picking tiles out of this "pool", instead of looping through the board looking
for the letter we need (by direct array access with some arithmetic on the
character ASCII codes). The subarrays allow us to construct a move array which
tracks the positions of the letters used on the board, so no information is
lost there. These move arrays use a different representation; the indices
correspond to the letter's position in the formed word, and the values are
the index in the board where the letter is found. This will actually allow for
more efficient move application, since we only need to loop through as many
tiles as are affected by the move.

The code to actually evaluate the moves in the list and find the best ones
still needs to be written, as well as some old code relevant only to the old
algorithm deleted. We can also implement the "vulnerable tile preference"
of the old algorithm in an easier, better way, simply by putting the
vulnerable tile positions at the end of the subarrays in the alphabet pool,
so they are always used first. There ARE a few caveats to this approach.
Most significantly, by design the new algorithm only generates one move for
every word in the TST, using the letters at the end of the subarrays. If we
use vulnerable tile preferencing, this will likely produce the best move for
any given word, but not always. We can remedy this by branching further when
we have multiple of the same letter in the pool, but we should ONLY branch
on vulnerable tiles as it doesn't matter which invulnerable tiles we use. This
will necessarily add to processing time (more the more duplicate letters there
are), but it may be worth it to ensure we gather all useful move permutations
of each word. Secondly, the "smart ordering" of "likely better" moves toward
the front of the list is lost, as we always traverse the TST in
middle-left-right order. However, we can preserve "extend-first" ordering by
adding the current node's move at the very end of every findMoves call, if
desired. This is likely unimportant, though, as alpha-beta pruning will quickly
eliminate any short words. When adding a second ply, it will probably be very
useful to first sort (perhaps partially or pseudo-) the moves by their
single-ply value (iterative-deepening style) for better pruning.