moveList generation with all combos added, BUGGY
Generating all combos seems to be working - size of move list for sample
game increased from ~7000 to ~77000, and processing time multiplied similarly.
However, results for the top 10 moves are identical, so somehow we're getting
(lots of) duplicates. This is somewhat encouraging, though, as this means
processing time could be significantly reduced IF we can remove these
duplicates.