Fixed combination bug
Was removing elements from beginning of array instead of end. Fixed, and
now it works beautifully. Running time is still the same, as we weren't
producing extra elements generally but rather as many combos of each word as
we should have had, but copies of one combo rather than different ones.