BUGS GALORE: blocking for initial words broken
The blocking is failing, returning false for both praiseworthy and orthopterist.
Stuff is going very strange with the initial findSibling call in block (gives a
node with the wrong letter) as well as in the node navigation in delete calls.
Need to add even more annotation and get to the bottom of it.