Currently infinite looping; DO NOT TEST
Presumably fixed the insertion problem, but now running is either infinite
looping or just taking forever. Started occurring somewhere between fixing
the creation of new left and right nodes in this.insert and creation of next.
Test with very small dictionary and print stuff in the loop to confirm whether
it's infinite.