TST bug fixed (MOSTLY)
Added a bunch of debugging prints, and ran with larger and larger dictionaries.
Now it works on the main dictionary, and quite fast (4 seconds), so it seems
the separation made a nice difference in load time. I'm not sure how I fixed
the infinite looping/slowness bug, though; the only actual change I made was
manually popping off the last element of the words list (an empty line due to
split), which I don't think would have done too much. Anyway, it's fixed now.
However, there are still bugs. The lookup for praiseworthy is returning false,
even though manual inspection of the TST.root object shows that it's in there.
Gotta figure that out next.