Fixed lookup bug
Lookup was going one past the last node of the word (because of loop always
going to next at end), so was actually returning endsWord of the next node
(or false if it didn't exist). So that's fixed now and setup and lookup seem
to be working great. When the timeout checking was removed, the dictionary
sets up even faster, around 2 seconds. Still need to finish transferring
delete and block, etc. over to the TST class.