IDDFS implemented and working
Almost entirely done with logic. Just need to add prefix blocking in TST's
block function, then it's all done. Maybe clean up the code a bit (remove
some printing, but also keep a few just to track the steps of execution in
the console), then start working on the interface! P.S. Forgot a pretty
important part - need to add evaluation for ending boards: set the value of
these to +/- 52.