Still having problems with form
Can't seem to stop the default behavior of the form reloading the page and
sending the access-code via GET; this only happens on the first submit, then
it works fine... also, the value of the input text field is coming out
undefined whatever I enter. No idea why either of these are happening after
much research.