Started open game animation
Had to use a strange setTimeout hack to get the loading page to show up
while the game is loading. Both .css and .show seemed to acting asynchronously,
so the only way was to put the rest of the code in a setTimeout(code, 0). That
worked well, but it's a bit hacky and possibly not super compatible. May want
to try using JQuery's .delay instead for better browser compatibly.