Added more game play page; changed IDDFS narrowing
Game play page not at all done, just had a sudden realization that the IDDFS
is shortening the list a bit too much each time by doing the square root of
the length of the last list, and overshortening could actually lead to worse
results by not considering all the important opponent responses at greater
depths. So I changed Game.js to shorten the list to the exact length to
guarantee that each level takes at most exactly as long as depth 0 - that is,
for n-ply lookahead, the length of the list considered is the nth root of the
length of the original list. Thus, even if pruning doesn't do ANYTHING, the
processing will be at most Math.pow(Math.pow(origLen, (1/n), n); i.e. origLen.