Fixed a few bugs
Still many more to fix. I think I've gotten new game creation/saving working
correctly, but there are still some issues. The new game I created isn't
updating the preview-board correctly when chosen, and the score is reported
incorrectly (0-25). Not sure if the game data was saved correctly and this is
just a display error, or if the game wasn't created/saved correctly.